\name{SnP500List}
\alias{SnP500List}
\docType{data}
\title{List of S&P500 Stocks in 2013}
\description{
List of company names corresponding to the ticker names of the stocks listed in S&P500 List in the year 2013.
}
\usage{data(SnP500List)}
\format{
  A data frame with 500 observations on the following variable.
  \describe{
    \item{\code{Company}}{Names of companies as character string corresponding to their ticker names as rowname.}
  }
}

\examples{
data(SnP500List)
head(SnP500List)
}
\keyword{datasets}
