\name{PortRisk-package}
\alias{PortRisk-package}
\alias{PortRisk}
\docType{package}
\title{
Portfolio Risk Analysis
}
\description{
Risk Attribution of a portfolio with Volatility Risk Analysis.
}
\details{
\tabular{ll}{
Package: \tab PortRisk\cr
Type: \tab Package\cr
Version: \tab 1.1.0\cr
Date: \tab 2015-10-31\cr
License: \tab GPL-2 | GPL-3\cr
Depends: \tab R (>= 3.0.0)\cr
Imports: \tab \code{zoo}, \code{tseries}, \code{MCMCpack}, \code{copula}, \code{MASS}
}
This package includes functions to compute the volatility risk attributes such as Volatility, Portfolio Volatility, MCTR, CCTR etc.
}
\author{
Sourish Das [aut, cre], Tamal Kanti Panja [aut].

Maintainer: Sourish Das <sourish.das@gmail.com>
}

% \references{
% ~~ Literature or other references for background information ~~
% }
% ~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
% ~~ documentation directory ~~
% \keyword{ package }
% \seealso{
% ~~ Optional links to other man pages, e.g. ~~
% ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
% }